backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
yfinance==0.2.50
borsapy==0.1.4
requests==2.32.3
pandas==2.2.3
numpy==2.1.3
httpx==0.27.2
holidays>=0.62
